requirements.txt
python-dotenv
streamlit>=1.32
requests>=2.31
python-dotenv>=1.0
openai>=1.30
python-docx>=1.0
pdfplumber>=0.11
PyPDF2>=3.0
